Model response bodies for all article routes
