+++ examples/realworld/conduit_core/src/api/articles/list_articles.rs
@@ -1,5 +1,7 @@
 use pavex::{extract::query::QueryParams, hyper::StatusCode};
 
+use crate::schemas::Article;
+
 #[derive(Debug, serde::Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ListArticles {
@@ -25,30 +27,6 @@ fn default_offset() -> u64 {
 pub struct ListArticlesResponse {
     pub articles: Vec<Article>,
     pub articles_count: u64,
-}
-
-#[derive(Debug, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Article {
-    pub slug: String,
-    pub title: String,
-    pub description: String,
-    pub body: String,
-    pub tag_list: Vec<String>,
-    pub created_at: String,
-    pub updated_at: String,
-    pub favorited: bool,
-    pub favorites_count: u64,
-    pub author: Profile,
-}
-
-#[derive(Debug, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Profile {
-    pub username: String,
-    pub bio: String,
-    pub image: String,
-    pub following: bool,
 }
 
 pub fn list_articles(_params: QueryParams<ListArticles>) -> StatusCode {
