+++ examples/realworld/conduit_core/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod api;
+pub mod schemas;
